moved some unit testing from header to articleList and added basic unitt tests to marsRover as well
