+++ client/src/components/MarsRover/index.js
@@ -1,0 +1,18 @@
+import React from 'react';
+
+const MarsRover = () => {
+  return (
+    <div>
+      <h1>Hello World!</h1>
+    </div>
+  );
+};
+
+export default MarsRover;
+
+
+
+
+
+
+
